Move Names and DisplayName into the contact package
