+++ android_mobile/src/main/java/com/alexstyl/specialdates/contact/Names.java
@@ -1,4 +1,4 @@
-package com.alexstyl.specialdates;
+package com.alexstyl.specialdates.contact;
 
 import java.util.Arrays;
 import java.util.Iterator;
