+++ android_mobile/src/main/java/com/alexstyl/specialdates/upcoming/widget/list/CircularAvatarFactory.java
@@ -9,7 +9,7 @@ import android.graphics.Typeface;
 import android.support.annotation.Px;
 
 import com.alexstyl.resources.ColorResources;
-import com.alexstyl.specialdates.DisplayName;
+import com.alexstyl.specialdates.contact.DisplayName;
 import com.alexstyl.specialdates.Optional;
 import com.alexstyl.specialdates.R;
 import com.alexstyl.specialdates.contact.Contact;
