+++ android_mobile/src/test/java/com/alexstyl/specialdates/contact/ContactTest.java
@@ -2,8 +2,6 @@ package com.alexstyl.specialdates.contact;
 
 import android.net.Uri;
 
-import com.alexstyl.specialdates.DisplayName;
-
 import org.junit.Before;
 import org.junit.Test;
