+++ android_mobile/src/test/java/com/alexstyl/specialdates/date/ContactEventTest.java
@@ -1,7 +1,7 @@
 package com.alexstyl.specialdates.date;
 
 import com.alexstyl.resources.StringResources;
-import com.alexstyl.specialdates.DisplayName;
+import com.alexstyl.specialdates.contact.DisplayName;
 import com.alexstyl.specialdates.Optional;
 import com.alexstyl.specialdates.R;
 import com.alexstyl.specialdates.TestContact;
