+++ android_mobile/src/test/java/com/alexstyl/specialdates/events/ContactActionTest.java
@@ -1,6 +1,6 @@
 package com.alexstyl.specialdates.events;
 
-import com.alexstyl.specialdates.DisplayName;
+import com.alexstyl.specialdates.contact.DisplayName;
 import com.alexstyl.specialdates.Optional;
 import com.alexstyl.specialdates.TestContact;
 import com.alexstyl.specialdates.contact.Contact;
