+++ android_mobile/src/test/java/com/alexstyl/specialdates/facebook/friendimport/FacebookContactFactoryTest.java
@@ -1,6 +1,6 @@
 package com.alexstyl.specialdates.facebook.friendimport;
 
-import com.alexstyl.specialdates.DisplayName;
+import com.alexstyl.specialdates.contact.DisplayName;
 import com.alexstyl.specialdates.contact.Contact;
 import com.alexstyl.specialdates.date.ContactEvent;
 import com.alexstyl.specialdates.date.DateParseException;
